[Decomp][Group] Cohesion split: extract member state to GroupMemberState.cpp
